% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver.R
\name{tr_series_acf}
\alias{tr_series_acf}
\title{Correlograma: a autocorrelação em cada defasagem.}
\usage{
tr_series_acf(
  serie,
  defasagens = 0L,
  aspecto = "16:9",
  tema = "padrão",
  titulo = "",
  rotulo_x = "",
  rotulo_y = "",
  legenda = "direita"
)
}
\description{
Correlograma: a autocorrelação em cada defasagem.
}
