% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ver.R
\name{tr_series_pacf}
\alias{tr_series_pacf}
\title{Autocorrelação parcial: a correlação com a defasagem k, descontadas as
intermediárias.}
\usage{
tr_series_pacf(
  serie,
  defasagens = 0L,
  aspecto = "16:9",
  tema = "padrão",
  titulo = "",
  rotulo_x = "",
  rotulo_y = "",
  legenda = "direita"
)
}
\description{
Autocorrelação parcial: a correlação com a defasagem k, descontadas as
intermediárias.
}
