% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada.R
\name{tr_series_pick}
\alias{tr_series_pick}
\title{Uma coluna da série múltipla, de volta ao mundo univariado.}
\usage{
tr_series_pick(series, variavel = "")
}
\description{
Uma coluna da série múltipla, de volta ao mundo univariado.
}
