% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariada.R
\name{tr_series_plot_mts}
\alias{tr_series_plot_mts}
\title{Uma faixa por série, eixo y livre: níveis diferentes não se espremem.}
\usage{
tr_series_plot_mts(
  serie,
  aspecto = "16:9",
  tema = "padrão",
  titulo = "",
  rotulo_x = "",
  rotulo_y = "",
  legenda = "direita"
)
}
\description{
Uma faixa por série, eixo y livre: níveis diferentes não se espremem.
}
